requirements.txt
bcrypt
pyjwt
pytest
sqlite3
cryptography
